Moved out of internal pkg & initial err has \n
